Use portable format specifiers: Case %ld - localized - smsg.
Fix uses of localized "%ld" within smsg():
- Replace "%ld" with "%" PRId64.
- Cast corresponding argument to (int64_t).

+++ src/ex_cmds.c
@@ -756,7 +756,7 @@ int do_move(linenr_T line1, linenr_T line2, linenr_T dest)
     if (num_lines == 1)
       MSG(_("1 line moved"));
     else
-      smsg((char_u *)_("%ld lines moved"), num_lines);
+      smsg((char_u *)_("%" PRId64 " lines moved"), (int64_t)num_lines);
   }
 
   /*
